src/components/fondant_llama_write_to_vector_db/requirements.txt
weaviate-client
llama-index
fondant[docker]==0.6.2